+++ pkg/event_store_on_dynamodb.go
@@ -1,4 +1,4 @@
-package event_store_adapter_go
+package pkg
 
 import (
 	"context"
